refining action buttons
