+++ src/common/ActionButtons/ActionButtons.module.css
@@ -3,6 +3,6 @@
 	display: flex;
 	flex-direction: row;
 	flex-wrap: wrap;
-	gap: 10px 6px;
+	gap: 8px;
 	max-width: 295px;
 }
